perf: pre compute lagrange coset basis poly once
