+++ internal/generator/backend/template/zkpschemes/plonk/plonk.marshal.go.tmpl
@@ -117,6 +117,8 @@ func (pk *ProvingKey) WriteTo(w io.Writer) (n int64, err error) {
 		}
 	}
 
+	pk.computeLagrangeCosetPolys()
+
 	return n + enc.BytesWritten(), nil
 }
